NFC: Split common crate into smaller modules
